database: extract storage info
